move H160Wrapper to shared
