+++ orderbook/src/api/get_fee_info.rs
@@ -1,11 +1,11 @@
 use crate::fee::{EthAwareMinFeeCalculator, MinFeeCalculating, MinFeeCalculationError};
 
-use super::H160Wrapper;
 use anyhow::Result;
 use chrono::{DateTime, Utc};
 use model::{order::OrderKind, u256_decimal};
 use primitive_types::{H160, U256};
 use serde::{Deserialize, Serialize};
+use shared::H160Wrapper;
 use std::convert::Infallible;
 use std::sync::Arc;
 use warp::{hyper::StatusCode, reply, Filter, Rejection, Reply};
