+++ shared/Cargo.toml
@@ -15,6 +15,7 @@ derivative = "2.2"
 ethcontract = { version = "0.12.2", default-features = false }
 futures = "0.3"
 gas-estimation = { git = "https://github.com/gnosis/gp-gas-estimation.git", tag = "v0.2.0", features = ["web3_"] }
+hex = { version = "0.4", default-features = false }
 hex-literal = "0.3"
 itertools = "0.10"
 lazy_static = "1.4.0"
