+++ shared/src/lib.rs
@@ -1,3 +1,9 @@
+use ethcontract::H160;
+use hex::{FromHex, FromHexError};
+use model::h160_hexadecimal;
+use serde::Deserialize;
+use std::str::FromStr;
+
 #[macro_use]
 pub mod macros;
 
@@ -31,3 +37,15 @@ pub mod transport;
 pub mod web3_traits;
 
 pub type Web3 = web3::Web3<transport::MetricTransport<crate::http_transport::HttpTransport>>;
+
+/// Wraps H160 with FromStr and Deserialize that can handle a `0x` prefix.
+#[derive(Deserialize)]
+#[serde(transparent)]
+pub struct H160Wrapper(#[serde(with = "h160_hexadecimal")] pub H160);
+impl FromStr for H160Wrapper {
+    type Err = FromHexError;
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let s = s.strip_prefix("0x").unwrap_or(s);
+        Ok(H160Wrapper(H160(FromHex::from_hex(s)?)))
+    }
+}
